move util function to htk.utils.request
